#6.12 Refactoring
